Tests/LibWeb/Text/expected/Editing/home-end-with-floated-first-letter.txt

End: 22
Home: 0
Down: 1
Home: 0
End: 22
